requirements.txt
customtkinter == 5.2.0
tk == 0.1.0
pycryptodome == 3.19.0
sympy == 1.12
pybase64 == 1.3.0